+++ src/basic/entities/notification/types.ts
