+++ src/basic/entities/notification/ui/NotificationItem.tsx
